dcsm_agent_sdk_python/requirements.txt
grpcio
grpcio-tools
protobuf
cachetools
